Write README. Template conf-files. Touch up script

+++ do_incremental_rsync.sh
@@ -1,28 +1,36 @@
 #!/bin/bash
 
 # Usage:
-# - mount your backup target (NAS or other remote disk) somewhere in the filesystem.
-# - edit CONFIGURATION below
-# - edit backup_exclude.conf and keep in the same directory as this script
-# - test on some smaller directory. run with -v to see verbose rsync output
-
+#  do_incremental_rsync.sh /some/directory /and/some/other/directory
+#  do_incremental_rsync.sh                (will backup /)
+#
+#  do_incremental_rsync.sh -v /home /     (-v = print verbose rsync output)
+#
+# rsync uses --one-file-system, so if you have several filesystems under / you need to supply them as separate arguments to the script
+#
 # More information in the README or http://ekenberg.github.io/linux-timemachine/
 
-# run without arguments to backup /, or supply list of directories to backup
-# rsync uses --one-file-system, so if you have several filesystems under / you need to supply them all as arguments to the script
-
-## CONFIGURATION:
-BACKUP_BASE=/path/to/directory/where/backups/are/stored
-BACKUP_NAME=your-computername-backup
-## END CONFIGURATION
+function die {
+    echo >&2 $@
+    exit 1
+}
 
 # Go to directory of this script
 cd "$( dirname "$0" )"
 
 if [ `id -u` != "0" ]; then
-    echo "You need to be root (sudo) to run system backups"
+    echo "You need to be root (or sudo) to run system backups"
     exit 1
 fi
+
+# Include configuration
+source backup.conf || die "You must supply backup.conf"
+
+[ -z "${BACKUP_BASE}" ] && die "BACKUP_BASE not configured in backup.conf"
+[ -z "${BACKUP_NAME}" ] && die "BACKUP_NAME not configured in backup.conf"
+
+# check for exclude-file
+[ -f backup_exclude.conf ] || die "You must supply backup_exclude.conf (empty is ok)"
 
 VERBOSE_ARGS=""
 if [ "x$1" = "x-v" ]; then
@@ -42,10 +50,7 @@ NEW_BACKUP="${BACKUP_BASE}/${BACKUP_NAME}-${TODAY}"
 
 mkdir -p "$NEW_BACKUP"
 
-if [ ! -d "$NEW_BACKUP" ]; then
-    echo "No such directory: $NEW_BACKUP"
-    exit 1
-fi
+[ -d "$NEW_BACKUP" ] || die "No such directory: $NEW_BACKUP"    
 
 for backup_item in "${BACKUP_WHAT[@]}"; do
 	backup_item_real=`realpath "$backup_item" 2>/dev/null`
@@ -59,9 +64,7 @@ for backup_item in "${BACKUP_WHAT[@]}"; do
 	rsync $VERBOSE_ARGS -a --delete --relative --one-file-system --numeric-ids --exclude-from=backup_exclude.conf --link-dest="$CURRENT_BACKUP" "$backup_item_real" "$NEW_BACKUP"
 done
 
-# Update symbolic link to current
-if [ -h "$CURRENT_BACKUP" ]; then
-    rm -f "$CURRENT_BACKUP"
-    ln -s "$NEW_BACKUP" "$CURRENT_BACKUP"
-fi
+# Update soft link to current backup
+[ -h "$CURRENT_BACKUP" ] && rm -f "$CURRENT_BACKUP"
+ln -s "$NEW_BACKUP" "$CURRENT_BACKUP" || die "Cannot create soft link '$NEW_BACKUP' -> '$CURRENT_BACKUP'"
